Removed the circular import by moving, the variables CURRENT_SEASON and CURRENT_TEAMS into a new file called season.py, deleted the file  since was deprecated in favor of
